+++ src/database/schema.sql
@@ -1,3 +1,7 @@
+DROP TYPE IF EXISTS user_role CASCADE;
+
+CREATE TYPE user_role AS ENUM ('admin', 'user');
+
 DROP EXTENSION IF EXISTS pg_stat_statements CASCADE;
 
 DROP TABLE IF EXISTS loginuser CASCADE;
@@ -26,7 +30,7 @@ DROP TABLE IF EXISTS hashtag CASCADE;
 
 DROP TABLE IF EXISTS job CASCADE;
 
-DROP TABLE IF EXISTS FAQ CASCADE;
+DROP TABLE IF EXISTS faq CASCADE;
 
 DROP TABLE IF EXISTS listinglike CASCADE;
 
@@ -48,37 +52,38 @@ DROP TABLE IF EXISTS listinglocation CASCADE;
 
 DROP TABLE IF EXISTS organisationannouncement CASCADE;
 
-CREATE EXTENSION pg_stat_statements;
+CREATE EXTENSION IF NOT EXISTS pg_stat_statements;
 
 CREATE TABLE loginuser (
-	user_id VARCHAR,
+	user_id UUID,
 	first_name VARCHAR NOT NULL,
 	last_name VARCHAR,
 	email VARCHAR(320) UNIQUE NOT NULL,
 	password VARCHAR NOT NULL,
-	role VARCHAR NOT NULL CHECK (role IN ('admin', 'user')) DEFAULT 'user',
+	role user_role NOT NULL DEFAULT 'user',
 	is_activated BOOLEAN NOT NULL DEFAULT FALSE,
 	PRIMARY KEY (user_id)
 );
 
 CREATE TABLE pendinguser (
 	pending_user_id SERIAL,
-	user_id VARCHAR UNIQUE NOT NULL,
+	user_id UUID UNIQUE NOT NULL,
 	token VARCHAR UNIQUE NOT NULL,
 	PRIMARY KEY (pending_user_id),
 	FOREIGN KEY (user_id) REFERENCES loginuser ON DELETE CASCADE
 );
 
 CREATE TABLE forgetpassworduser (
-	forgetpass_user_id SERIAL,
+	forget_password_user_id SERIAL,
 	email VARCHAR(320) UNIQUE NOT NULL,
 	token VARCHAR UNIQUE NOT NULL,
 	expiry TIMESTAMPTZ NOT NULL DEFAULT NOW(),
-	PRIMARY KEY (forgetpass_user_id)
+	PRIMARY KEY (forget_password_user_id),
+	FOREIGN KEY (email) REFERENCES loginuser (email) ON DELETE CASCADE
 );
 
 CREATE TABLE profile (
-	user_id VARCHAR,
+	user_id UUID,
 	nickname VARCHAR NOT NULL,
 	profile_picture VARCHAR,
 	about TEXT,
@@ -105,7 +110,7 @@ CREATE TABLE organisation (
 	phone VARCHAR,
 	email VARCHAR(320),
 	address VARCHAR(320),
-	owned_by VARCHAR,
+	owned_by UUID,
 	locations VARCHAR[],
 	story TEXT,
 	facebook_link VARCHAR,
@@ -116,14 +121,15 @@ CREATE TABLE organisation (
 	additional_photos VARCHAR[],
 	is_verified BOOLEAN NOT NULL DEFAULT FALSE,
 	created_on TIMESTAMPTZ NOT NULL DEFAULT NOW(),
-	deleted_on TIMESTAMPTZ DEFAULT NULL,
-	PRIMARY KEY (organisation_id)
+	deleted_on TIMESTAMPTZ,
+	PRIMARY KEY (organisation_id),
+	FOREIGN KEY (owned_by) REFERENCES loginuser (user_id) ON DELETE SET NULL
 );
 
 CREATE TABLE membership (
 	membership_id SERIAL,
 	organisation_id UUID NOT NULL,
-	user_id VARCHAR NOT NULL,
+	user_id UUID NOT NULL,
 	is_owner BOOLEAN NOT NULL DEFAULT FALSE,
 	joined_on TIMESTAMPTZ NOT NULL DEFAULT NOW(),
 	PRIMARY KEY (membership_id),
@@ -150,8 +156,8 @@ CREATE TABLE category (
 );
 
 CREATE TABLE listing (
-	listing_id VARCHAR,
-	created_by VARCHAR,
+	listing_id UUID,
+	created_by UUID,
 	title VARCHAR NOT NULL,
 	category VARCHAR,
 	about TEXT,
@@ -168,7 +174,7 @@ CREATE TABLE listing (
 	end_date TIMESTAMPTZ,
 	created_on TIMESTAMPTZ NOT NULL DEFAULT NOW(),
 	updated_on TIMESTAMPTZ NOT NULL DEFAULT NOW(),
-	deleted_on TIMESTAMPTZ DEFAULT NULL,
+	deleted_on TIMESTAMPTZ,
 	PRIMARY KEY (listing_id),
 	FOREIGN KEY (category) REFERENCES category (category_name) ON DELETE SET NULL,
 	FOREIGN KEY (created_by) REFERENCES loginuser (user_id) ON DELETE SET NULL
@@ -176,7 +182,7 @@ CREATE TABLE listing (
 
 CREATE TABLE listingorganisation (
 	listing_organisation_id SERIAL,
-	listing_id VARCHAR NOT NULL,
+	listing_id UUID NOT NULL,
 	organisation_id UUID NOT NULL,
 	PRIMARY KEY (listing_organisation_id),
 	UNIQUE (listing_id, organisation_id),
@@ -185,7 +191,7 @@ CREATE TABLE listingorganisation (
 );
 
 CREATE TABLE listingstory (
-	listing_id VARCHAR,
+	listing_id UUID,
 	overview TEXT,
 	problem TEXT,
 	solution TEXT,
@@ -196,7 +202,7 @@ CREATE TABLE listingstory (
 
 CREATE TABLE hashtag (
 	hashtag_id SERIAL,
-	listing_id VARCHAR NOT NULL,
+	listing_id UUID NOT NULL,
 	tag VARCHAR NOT NULL,
 	PRIMARY KEY (hashtag_id),
 	UNIQUE (listing_id, tag),
@@ -206,17 +212,17 @@ CREATE TABLE hashtag (
 /* Jobs for a particular listing */
 CREATE TABLE job (
 	job_id SERIAL,
-	listing_id VARCHAR NOT NULL,
+	listing_id UUID NOT NULL,
 	job_title VARCHAR NOT NULL,
 	job_description TEXT,
-	deleted_on TIMESTAMPTZ DEFAULT NULL,
+	deleted_on TIMESTAMPTZ,
 	PRIMARY KEY (job_id),
 	FOREIGN KEY (listing_id) REFERENCES listing ON DELETE CASCADE
 );
 
-CREATE TABLE FAQ (
+CREATE TABLE faq (
 	faq_id SERIAL,
-	listing_id VARCHAR NOT NULL,
+	listing_id UUID NOT NULL,
 	question TEXT NOT NULL,
 	answer TEXT,
 	PRIMARY KEY (faq_id),
@@ -225,8 +231,8 @@ CREATE TABLE FAQ (
 
 CREATE TABLE listinglike (
 	like_id SERIAL,
-	user_id VARCHAR NOT NULL,
-	listing_id VARCHAR NOT NULL,
+	user_id UUID NOT NULL,
+	listing_id UUID NOT NULL,
 	PRIMARY KEY (like_id),
 	UNIQUE (user_id, listing_id),
 	FOREIGN KEY (user_id) REFERENCES loginuser ON DELETE CASCADE,
@@ -236,17 +242,17 @@ CREATE TABLE listinglike (
 CREATE TABLE organisationlike (
 	organisation_like_id SERIAL,
 	organisation_id UUID NOT NULL,
-	user_id VARCHAR NOT NULL,
+	user_id UUID NOT NULL,
 	PRIMARY KEY (organisation_like_id),
 	UNIQUE (organisation_id, user_id),
-	FOREIGN KEY (user_id) REFERENCES Users ON DELETE CASCADE,
-	FOREIGN KEY (organisation_id) REFERENCES Organisations ON DELETE CASCADE
+	FOREIGN KEY (user_id) REFERENCES loginuser ON DELETE CASCADE,
+	FOREIGN KEY (organisation_id) REFERENCES organisation ON DELETE CASCADE
 );
 
 CREATE TABLE listingadmin (
 	listing_admin_id SERIAL,
-	user_id VARCHAR NOT NULL,
-	listing_id VARCHAR NOT NULL,
+	user_id UUID NOT NULL,
+	listing_id UUID NOT NULL,
 	PRIMARY KEY (listing_admin_id),
 	UNIQUE (user_id, listing_id),
 	FOREIGN KEY (user_id) REFERENCES loginuser ON DELETE CASCADE,
@@ -255,8 +261,8 @@ CREATE TABLE listingadmin (
 
 CREATE TABLE participant (
 	participant_id SERIAL,
-	listing_id VARCHAR NOT NULL,
-	user_id VARCHAR NOT NULL,
+	listing_id UUID NOT NULL,
+	user_id UUID NOT NULL,
 	joined_on TIMESTAMPTZ NOT NULL DEFAULT NOW(),
 	end_on TIMESTAMPTZ,
 	PRIMARY KEY (participant_id),
@@ -267,8 +273,8 @@ CREATE TABLE participant (
 
 CREATE TABLE milestone (
 	milestone_id SERIAL,
-	listing_id VARCHAR NOT NULL,
-	description TEXT NOT NULL,
+	listing_id UUID NOT NULL,
+	description TEXT,
 	date TIMESTAMPTZ,
 	PRIMARY KEY (milestone_id),
 	FOREIGN KEY (listing_id) REFERENCES listing ON DELETE CASCADE
@@ -276,8 +282,8 @@ CREATE TABLE milestone (
 
 CREATE TABLE listingupdate (
 	listing_update_id SERIAL,
-	listing_id VARCHAR NOT NULL,
-	description TEXT NOT NULL,
+	listing_id UUID NOT NULL,
+	description TEXT,
 	pics VARCHAR[],
 	created_on TIMESTAMPTZ NOT NULL DEFAULT NOW(),
 	updated_on TIMESTAMPTZ NOT NULL DEFAULT NOW(),
@@ -287,8 +293,8 @@ CREATE TABLE listingupdate (
 
 CREATE TABLE listingcomment (
 	listing_comment_id SERIAL,
-	listing_id VARCHAR,
-	user_id VARCHAR,
+	listing_id UUID,
+	user_id UUID,
 	COMMENT TEXT,
 	reply_to_id INTEGER,
 	created_on TIMESTAMPTZ NOT NULL DEFAULT NOW(),
@@ -302,14 +308,14 @@ CREATE TABLE listingcomment (
 
 CREATE TABLE location (
 	location_id SERIAL,
-	location VARCHAR,
+	location_name VARCHAR UNIQUE NOT NULL,
 	zone VARCHAR,
 	PRIMARY KEY (location_id)
 );
 
 CREATE TABLE listinglocation (
 	listing_location_id SERIAL,
-	listing_id VARCHAR NOT NULL,
+	listing_id UUID NOT NULL,
 	location_id INTEGER NOT NULL,
 	PRIMARY KEY (listing_location_id),
 	UNIQUE (listing_id, location_id),
@@ -320,10 +326,10 @@ CREATE TABLE listinglocation (
 CREATE TABLE organisationannouncement (
 	announcement_id SERIAL,
 	organisation_id UUID NOT NULL,
-	ANNOUNCEMENT TEXT NOT NULL,
+	announcement TEXT NOT NULL,
 	created_on TIMESTAMPTZ NOT NULL DEFAULT NOW(),
 	updated_on TIMESTAMPTZ NOT NULL DEFAULT NOW(),
 	deleted_on TIMESTAMPTZ,
 	PRIMARY KEY (announcement_id),
-	FOREIGN KEY (organisation_id) REFERENCES Organisations ON DELETE CASCADE
+	FOREIGN KEY (organisation_id) REFERENCES organisation ON DELETE CASCADE
 );